Refactoring
Moved the function and added the file.
Added tests and test data.
