+++ oviewer/document.go
@@ -124,6 +124,8 @@ func (m *Document) ReadFile(fileName string) error {
 	return nil
 }
 
+// Close closes the File.
+// Record the last read position.
 func (m *Document) Close() error {
 	pos, err := m.file.Seek(0, io.SeekCurrent)
 	if err != nil {
@@ -136,11 +138,14 @@ func (m *Document) Close() error {
 	return nil
 }
 
+// reOpenRead reopens and reads the file.
+// Seek to the position where the file was closed, and then read.
 func (m *Document) reOpenRead() error {
 	r, err := os.Open(m.FileName)
 	if err != nil {
 		return err
 	}
+	m.file = r
 	atomic.StoreInt32(&m.eof, 0)
 
 	_, err = r.Seek(m.offset, io.SeekStart)
